extension ImageSource and ImageLocation extracted
